# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MusicBox API Documentation",
        "description": "API Documentation for MusicBox",
        "contact": {
            "email": "your-email@domain.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/artists": {
            "get": {
                "tags": [
                    "Artists"
                ],
                "summary": "Get list of artists",
                "operationId": "392d5b22a13c0e6acda09059973b671b",
                "parameters": [
                    {
                        "name": "genre",
                        "in": "query",
                        "description": "Filter artists by genre",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "artists": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Artist Name"
                                                            },
                                                            "genre": {
                                                                "type": "string",
                                                                "example": "Rock"
                                                            },
                                                            "country": {
                                                                "type": "string",
                                                                "example": "USA"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Artists Fetched Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Artists"
                ],
                "summary": "Create a new artist",
                "operationId": "02d849b7e76cdecf89814f7cf5ad17e3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "genre",
                                    "country"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "New Artist"
                                    },
                                    "genre": {
                                        "type": "string",
                                        "example": "Rock"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "France"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Artist created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "artist": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "New Artist"
                                                },
                                                "genre": {
                                                    "type": "string",
                                                    "example": "Rock"
                                                },
                                                "country": {
                                                    "type": "string",
                                                    "example": "France"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Artist Created Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/artists/{id}": {
            "get": {
                "tags": [
                    "Artists"
                ],
                "summary": "Get artist by ID with their albums and songs",
                "operationId": "72409b143e96ba16780ec0cf7bf56b62",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Artist ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "artist": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Artist Name"
                                                },
                                                "genre": {
                                                    "type": "string",
                                                    "example": "Rock"
                                                },
                                                "country": {
                                                    "type": "string",
                                                    "example": "USA"
                                                },
                                                "albums": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "title": {
                                                                "type": "string",
                                                                "example": "Album Title"
                                                            },
                                                            "songs": {
                                                                "type": "array",
                                                                "items": {
                                                                    "properties": {
                                                                        "id": {
                                                                            "type": "integer",
                                                                            "example": 1
                                                                        },
                                                                        "title": {
                                                                            "type": "string",
                                                                            "example": "Song Title"
                                                                        }
                                                                    },
                                                                    "type": "object"
                                                                }
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Artist Fetched Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Artist not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Artists"
                ],
                "summary": "Update existing artist",
                "operationId": "5384c3d91429ef750acd6efe62cddef3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Artist ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "genre",
                                    "country"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Artist Name"
                                    },
                                    "genre": {
                                        "type": "string",
                                        "example": "Jazz"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "France"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Artist updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "artist": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Artist Updated Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Artist not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Artists"
                ],
                "summary": "Delete an artist",
                "operationId": "8b4067da292e7b4c226ffb95ecbe82e9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Artist ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Artist deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Artist Deleted Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Artist not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}